app/fabric-model-readiness/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fabric-model-readiness"
version = "0.1.0"
description = "Multi-agent system that analyzes Power BI semantic models for Microsoft Prep for AI best practices"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
scout = "scout.cli:app"
enforcer = "enforcer.cli:app"
historian = "historian.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/scout", "src/enforcer", "src/historian", "src/shared"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
